requirements.txt
Flask
pymongo
playwright
python-dotenv
Flask-Cors
gunicorn